add support for bankdetail & credit cards

+++ Shared/Dtos/BankDetail.cs
@@ -4,13 +4,13 @@ public class BankDetail
 {
     public string BankName { get; private set; }
         
-    public int Iban { get; private set; }
+    public double Iban { get; private set; }
         
     public string Username { get; private set; }
         
     public string Password { get; private set; }
 
-    public BankDetail(string bankName, int iban, string username, string password)
+    public BankDetail(string bankName, double iban, string username, string password)
     {
         BankName = bankName;
         Iban = iban;
