+++ Shared/Dtos/CreditCard.cs
@@ -6,15 +6,15 @@ public class CreditCard
 {
     public string Issuer { get; private set; }
         
-    public int Number { get; private set; }
+    public double Number { get; private set; }
         
     public DateTime ExpiryDate { get; private set; }
         
-    public int Cvv { get; private set; }
+    public double Cvv { get; private set; }
         
     public string BillingAddress { get; private set; }
 
-    public CreditCard(string issuer, int number, DateTime expiryDate, int cvv, string billingAddress)
+    public CreditCard(string issuer, double number, DateTime expiryDate, double cvv, string billingAddress)
     {
         Issuer = issuer;
         Number = number;
